requirements.txt
pillow
windnd
